Fix order in loading script. Update babel to latest version.
Removed async keyword using in index.html for loading
test-component.v.html as it's better to wait for the page to fully load
first (thus webcomponents-lite.min.js).

Updated babel to latest version. This means removing babel NPM package,
and use babel-core, and babel-cli instead. Also updated build.sh script
to work with cli from babel-cli.

Slightly modified karma config file but it doesn't make any impact as
Jasmine test suite is still not working. Issue 2 of karma-polymer is
fixed, and thus this project updated it to version 0.1.5 included with
that fix. But now it introduces with a new problem. It is still in
discussion and investigation.